fix(commands): Ask for password in backup and copy command if it is missing (#1061)
A user discovered that for some commands they are not being asked for a
password, although they should. This implements a fix for the backup and
copy command so they consistently call the same functionality that is
being used for the other commands. Hence, asking for the password, if
it's missing.