feat(ux): Use human-panic to print better error messages in case rustic panics (#1065)
human-panic is in so far valuable, that it can help us to better
diagnose panics and errors in rustic as it prints better messages and
even writes a log file with the stacktrace, that users can attach to
their issues.

human-panic only displays a human-friendly panic message in release
mode.