chore: Remove self-updater from default features
The self-updater feature is problematic for many different distribution
channels including distro packages, direct `cargo install`, various
package managers and containers, etc. It will even goof up folks
installing from source builds they made themselves. Either the binary
does not have permission and should not be trying to monkey with the
system installed binaries or the rest of the things besides the binary
will end up out of sync ith packaging.

Distros have their own update mechanisms that shouldn't be tampered with
by every app that comes along. This means distro have to build with
--no-default-features for apps that include self-updaters by default,
but that also means we have to maintain a list of features we do want.
This is not only tedious it is error-prone because there is a very good
chance of new features getting overlooked when doing version bumps.

Additionally many other install workflows should be updating using those
workflows, not being tampered with internally this way.

The binaries posted on releases that people might download and place
manually are good candidates for enabling this feature.