warn: warn about RUSTIC_APP.config() vs self usage
In commands that have configurations from
the ENV and the TOML (on top of CLI),
a confusion between RUSTIC_APP.config()
and &self (see `inner_run` impls) could
lead to parts of the user configured options
to be ignored.

For now we warn about this in the code.
But some design change should be decided
in order to avoid these risks in the future.