feat(interactive)!: Add interactive diff / overwork diff output (#1480)
Adds the possibility to compare two snapshots interactively.
From the interactive snapshots view,  press `D` to get to the diff view.
If two snapshots are marked, this diffs the two marked snapshots. If one
snapshot is marked or selected, this diffs with its parent (if there is
one).
Alternatively, you can use `rustic diff -i` to select snapshots/paths
directly.

The PR also allows to recursively compute added/removed space for dirs
and subdirs using `s` (like in #1479)

This PR is breaking as it also changes the output format of the normal
`diff` command.