fix(metrics): use short-lived local registry (#1575)
Previously, this was using the global registry (via register_gauge!),
which doesn't support re-registering metrics in this way, so only the
first call to `PrometheusExporter::push_metrics` would succeed, while
any later calls would return a (nested) `prometheus::Error::AlreadyReg`.

This resulted in only the metrics for the first snapshot being pushed,
while the metrics for subsequent snapshots errored with a somewhat vague
error (the interior messages are not shown):

```
[WARN] error pushing metrics: pushing prometheus metrics
```

This PR more closely matches the behavior of the OpenTelemetry backend,
which similarly only pushes the immediately-passed (as an argument)
metrics, rather than including other metrics previously registered
globally, as was the behavior with `prometheus::gather`.

There are likely less allocation-happy ways to solve this, but at the
push frequency of rustic it doesn't matter.

cc @Ekleog as the original author of this module