fix: Fix typos using the typos tool (#1590)
The [typos](https://github.com/crate-ci/typos) tool checks for common
typos in a codebase, and optionally fixes them if a single option is
possible.

I’ve fixed all the issues it found in rustic, ignored the two words it
considered wrong but were actually correct (ratatui and tpe), and added
a CI job to avoid reintroducing typos in the future.