feat(interactive): show summary in snapshots, ls and diff views (#1615)
When summaries are computed using the `s` shortcut in the ls and diff
view, they are now also displayed.
Moreover a `S` shortcut has been added in the ls, diff and snapshot view
to show summary for the currently selected entry (and all selected
snapshots in the snapshots view).

In the summary number of files, dirs, total (restore) size are shown as
well as blobs, deduplicated size, size on-repo and the compression
ratio.